backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
h11==0.16.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
MarkupSafe==3.0.3
pycparser==2.23
python-engineio==4.12.3
python-socketio==5.15.0
simple-websocket==1.1.0
typing_extensions==4.15.0
Werkzeug==3.1.4
wsproto==1.2.0
zipp==3.23.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
eventlet==0.36.1
